requirements.txt
Flask==0.11.1
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
marshmallow==2.10.4
nose==1.3.7
psycopg2==2.6.2
py-bcrypt==0.4
python_jwt==2.0.1
SQLAlchemy==1.1.4
